fix according to the review comments
