--- modules/video/multiple_object_tracking/jde_darknet53_30e_1088x608/config/_base_/jde_reader_1088x608.yml
@@ -1,57 +1,0 @@
-worker_num: 2
-TrainReader:
-  sample_transforms:
-    - Decode: {}
-    - RGBReverse: {}
-    - AugmentHSV: {}
-    - LetterBoxResize: {target_size: [608, 1088]}
-    - MOTRandomAffine: {}
-    - RandomFlip: {}
-    - BboxXYXY2XYWH: {}
-    - NormalizeBox: {}
-    - NormalizeImage: {mean: [0, 0, 0], std: [1, 1, 1], is_scale: True}
-    - RGBReverse: {}
-    - Permute: {}
-  batch_transforms:
-    - Gt2JDETargetThres:
-        anchor_masks: [[0, 1, 2, 3], [4, 5, 6, 7], [8, 9, 10, 11]]
-        anchors: [[[128,384], [180,540], [256,640], [512,640]],
-                  [[32,96], [45,135], [64,192], [90,271]],
-                  [[8,24], [11,34], [16,48], [23,68]]]
-        downsample_ratios: [32, 16, 8]
-        ide_thresh: 0.5
-        fg_thresh: 0.5
-        bg_thresh: 0.4
-  batch_size: 4
-  shuffle: true
-  drop_last: true
-  use_shared_memory: true
-
-
-EvalMOTReader:
-  sample_transforms:
-    - Decode: {}
-    - LetterBoxResize: {target_size: [608, 1088]}
-    - NormalizeImage: {mean: [0, 0, 0], std: [1, 1, 1], is_scale: True}
-    - Permute: {}
-  batch_size: 1
-
-
-TestMOTReader:
-  inputs_def:
-    image_shape: [3, 608, 1088]
-  sample_transforms:
-    - Decode: {}
-    - LetterBoxResize: {target_size: [608, 1088]}
-    - NormalizeImage: {mean: [0, 0, 0], std: [1, 1, 1], is_scale: True}
-    - Permute: {}
-  batch_size: 1
-
-
-MOTVideoStreamReader:
-  sample_transforms:
-    - Decode: {}
-    - LetterBoxResize: {target_size: [608, 1088]}
-    - NormalizeImage: {mean: [0, 0, 0], std: [1, 1, 1], is_scale: True}
-    - Permute: {}
-  batch_size: 1
