--- modules/video/multiple_object_tracking/jde_darknet53_30e_1088x608/config/_base_/optimizer_30e.yml
@@ -1,19 +1,0 @@
-epoch: 30
-
-LearningRate:
-  base_lr: 0.01
-  schedulers:
-  - !PiecewiseDecay
-    gamma: 0.1
-    milestones: [15, 22]
-    use_warmup: True
-  - !BurninWarmup
-    steps: 1000
-
-OptimizerBuilder:
-  optimizer:
-    momentum: 0.9
-    type: Momentum
-  regularizer:
-    factor: 0.0001
-    type: L2
