--- modules/video/multiple_object_tracking/jde_darknet53_30e_1088x608/config/_base_/runtime.yml
@@ -1,5 +1,0 @@
-use_gpu: true
-log_iter: 20
-save_dir: output
-snapshot_epoch: 1
-print_flops: false
