--- modules/video/multiple_object_tracking/jde_darknet53_30e_1088x608/config/jde_darknet53_30e_1088x608.yml
@@ -1,46 +1,0 @@
-_BASE_: [
-  '_base_/mot.yml',
-  '_base_/runtime.yml',
-  '_base_/optimizer_30e.yml',
-  '_base_/jde_darknet53.yml',
-  '_base_/jde_reader_1088x608.yml',
-]
-
-JDE:
-  detector: YOLOv3
-  reid: JDEEmbeddingHead
-  tracker: JDETracker
-
-YOLOv3:
-  backbone: DarkNet
-  neck: YOLOv3FPN
-  yolo_head: YOLOv3Head
-  post_process: JDEBBoxPostProcess
-  for_mot: True
-
-YOLOv3Head:
-  anchors: [[128,384], [180,540], [256,640], [512,640],
-            [32,96], [45,135], [64,192], [90,271],
-            [8,24], [11,34], [16,48], [23,68]]
-  anchor_masks: [[0, 1, 2, 3], [4, 5, 6, 7], [8, 9, 10, 11]]
-  loss: JDEDetectionLoss
-
-JDETracker:
-  det_thresh: 0.3
-  track_buffer: 30
-  min_box_area: 200
-  motion: KalmanFilter
-
-JDEBBoxPostProcess:
-  decode:
-    name: JDEBox
-    conf_thresh: 0.5
-    downsample_ratio: 32
-  nms:
-    name: MultiClassNMS
-    keep_top_k: 500
-    score_threshold: 0.01
-    nms_threshold: 0.4
-    nms_top_k: 2000
-    normalized: true
-    return_index: true
