--- modules/video/multiple_object_tracking/jde_darknet53_30e_1088x608/module.py
@@ -1,165 +1,0 @@
-# Copyright (c) 2020 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-import os
-import sys
-import signal
-
-parent_path = os.path.abspath(os.path.join(__file__, '..'))
-if parent_path not in sys.path:
-    sys.path.append(parent_path)
-
-# ignore warning log
-import warnings
-warnings.filterwarnings('ignore')
-
-import paddle
-from paddle.distributed import ParallelEnv
-
-from ppdet.core.workspace import load_config, merge_config
-from ppdet.utils.check import check_gpu, check_version, check_config
-from ppdet.utils.logger import setup_logger
-
-import paddlehub as hub
-from paddlehub.module.module import moduleinfo, serving, runnable
-
-import argparse
-from .tracker import StreamTracker
-
-logger = setup_logger('Predict')
-
-@moduleinfo(
-    name="jde_darknet53_30e_1088x608",
-    type="CV/multiple_object_tracking",
-    author="paddlepaddle",
-    author_email="",
-    summary="JDE is a joint detection and appearance embedding model for multiple object tracking.",
-    version="1.0.0")
-class JDETracker_1088x608(hub.Module):
-    def _initialize(self):
-        self.pretrained_model = os.path.join(self.directory, "jde_darknet53_30e_1088x608")
-
-    
-    def mot_predict(self, video_stream, output_dir='', visualization=True, draw_threshold=0.5, use_gpu=False, from_device=False):
-        cfg = load_config(os.path.join(self.directory, 'config', 'jde_darknet53_30e_1088x608.yml'))
-        check_config(cfg)
-
-        place = 'gpu:{}'.format(ParallelEnv().dev_id) if use_gpu else 'cpu'
-        place = paddle.set_device(place)
-    
-        paddle.disable_static() 
-        tracker = StreamTracker(cfg, mode='test')
-
-        # load weights
-        if cfg.architecture in ['DeepSORT']:
-            if cfg.det_weights != 'None':
-                tracker.load_weights_sde(cfg.det_weights, cfg.reid_weights)
-            else:
-                tracker.load_weights_sde(None, cfg.reid_weights)
-        else:
-            tracker.load_weights_jde(self.pretrained_model)
-        signal.signal(signal.SIGINT, self.signalhandler)
-        # inference
-        tracker.mot_predict(
-                    video_stream=video_stream,
-                    output_dir=output_dir,
-                    data_type='mot',
-                    model_type='JDE',
-                    visualization=visualization,
-                    draw_threshold=draw_threshold,
-                    from_device=from_device)
-
-    @serving
-    def serving_method(self, video_stream: str, **kwargs):
-        """
-        Run as a service.
-        """
-        pass
-
-    @runnable
-    def run_cmd(self, argvs: list):
-        """
-        Run as a command.
-        """
-        self.parser = argparse.ArgumentParser(
-            description="Run the {} module.".format(self.name),
-            prog='hub run {}'.format(self.name),
-            usage='%(prog)s',
-            add_help=True)
-
-        self.arg_input_group = self.parser.add_argument_group(
-            title="Input options", description="Input data. Required")
-        self.arg_config_group = self.parser.add_argument_group(
-            title="Config options", description="Run configuration for controlling module behavior, not required.")
-        self.add_module_config_arg()
-        self.add_module_input_arg()
-        self.args = self.parser.parse_args(argvs)
-        
-        self.mot_predict(
-            video_stream=self.args.video_stream,
-            output_dir=self.args.output_dir,
-            visualization=self.args.visualization,
-            draw_threshold=self.args.draw_threshold,
-            use_gpu = self.args.use_gpu,
-            from_device = self.args.from_device
-        )
-    
-    def signalhandler(self, signum, frame):
-        seq = os.path.splitext(os.path.basename(self.args.video_stream))[0]
-        save_dir = os.path.join(self.args.output_dir, 'mot_outputs',
-                                seq) if self.args.visualization else None
-        if self.args.visualization:
-            output_video_path = os.path.join(save_dir, '..',
-                                             '{}_vis.mp4'.format(seq))
-            cmd_str = 'ffmpeg -f image2 -i {}/%05d.jpg -vf "scale=trunc(iw/2)*2:trunc(ih/2)*2" {}'.format(
-                save_dir, output_video_path)
-            os.system(cmd_str)
-            print('Program Interrupted! Save video in {}'.format(output_video_path))
-            exit(0)
-        
-
-    def add_module_config_arg(self):
-        """
-        Add the command config options.
-        """
-        self.arg_config_group.add_argument(
-            '--use_gpu', action='store_true', help="use GPU or not")
-        
-        self.arg_config_group.add_argument(
-            '--from_device', action='store_true', help="=read from camera device")
-
-        self.arg_config_group.add_argument(
-            '--output_dir',
-            type=str,
-            default='mot_result',
-            help='Directory name for output tracking results.')
-        self.arg_config_group.add_argument(
-            '--visualization', 
-            type=bool, 
-            default=True, 
-            help="whether to save output as images.")
-        self.arg_config_group.add_argument(
-            "--draw_threshold",
-            type=float,
-            default=0.5,
-            help="Threshold to reserve the result for visualization.")
-
-    def add_module_input_arg(self):
-        """
-        Add the command input options.
-        """
-        self.arg_input_group.add_argument('--video_stream', type=str, help="path to video stream, can be a video file or stream device number.")
-         
-
-    
-
